Introduce metadata module
that has Versions and ListCommand class
